Added extra comments
