+++ MarsRover/DefaultSettings.cs
@@ -1,5 +1,3 @@
-using System.Collections.Generic;
-using MarsRover.Objectives;
 
 namespace MarsRover
 {
